+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## 1.0.1 - 2017-10-15
 
+### Added
+- Windows should display notification if everything is running well.
+
 ### Changed
 - Will try to display error in error box.